Updated jQuery & jQuery-UI scaffolding.
Made jQuery map conditional.  Added conditional unminified jQuery.
Changed jQuery-UI version.  Added missing jQuery-UI images & theme
stylesheet.  Refactored related templates.

+++ app/index.js
@@ -7,16 +7,6 @@ var http = require('http');
 var fs = require('fs');
 var postdata = require('postdata');
 
-var jqUiVerMap = { // jQuery version: jQuery-UI version.
-    '2.1.3': '1.11.2', 
-    '1.11.2': '1.11.2', 
-    '1.10.2': '1.10.4', 
-    '1.9.1': '1.9.2', 
-    '1.8.3': '1.8.24', 
-    '1.7.2': '1.8.24' 
-};
-var jqChoices = [];
-
 var TzeGenerator = yeoman.generators.Base.extend({
     initializing: function () {
         this.pkg = require('../package.json');
@@ -29,16 +19,6 @@ var TzeGenerator = yeoman.generators.Base.extend({
         if (!this.options['skip-welcome-message']) {
             console.log(yosay('WELCOME to the Tze Generator for Yeoman!\n' + 
                             'Let\'s get some basic project info/options set up...'));
-        }
-
-        var jqChoiceDefault = null;
-        var key = null;
-        for ( var jqVer in jqUiVerMap ) {
-            key = jqVer;
-            if ( !jqChoiceDefault ) {
-                jqChoiceDefault = key;
-            }
-            jqChoices.push({ name : key, value: key });
         }
 
         // Prompt for user inputs.
@@ -67,8 +47,15 @@ var TzeGenerator = yeoman.generators.Base.extend({
                 type: 'list',
                 name: 'jqVersion',
                 message: 'jQuery version',
-                choices: jqChoices,
-                default: jqChoiceDefault
+                choices: [
+                    { name: '2.1.3', value: '2.1.3' }, 
+                    { name: '1.11.2', value: '1.11.2' }, 
+                    { name: '1.10.2', value: '1.10.2'}, 
+                    { name: '1.9.1', value: '1.9.1'}, 
+                    { name: '1.8.3', value: '1.8.3'}, 
+                    { name: '1.7.2', value: '1.7.2'} 
+                ],
+                default: '2.1.3'
             },
             {
                 type: 'confirm',
@@ -105,6 +92,7 @@ var TzeGenerator = yeoman.generators.Base.extend({
         scaffoldFolders: function () {
             this.mkdir("src");
             this.mkdir("src/css");
+            this.mkdir("src/css/images");
             this.mkdir("src/fonts");
             this.mkdir("src/images");
             this.mkdir("src/sass");
@@ -121,7 +109,7 @@ var TzeGenerator = yeoman.generators.Base.extend({
             }
         },
         
-        downloadJQuery: function () {
+        downloadJQueryMin: function () {
             var done = this.async();
             var jqVrsn = this.jqVer;
             var dest = 'src/js/libs/jquery-' + jqVrsn + '.min.js';
@@ -131,14 +119,14 @@ var TzeGenerator = yeoman.generators.Base.extend({
                 response.pipe( file ).on( 'close', function () {
                      postdata( fs.readFileSync( dest ) );
                     if (!skipMsg) {
-                        console.log('[downloadJQuery] jQuery downloaded: ' + jqVrsn);
+                        console.log('[downloadJQueryMin] Minified jQuery downloaded: ' + jqVrsn);
                     }
                     done();
                 });
             })
             .on( 'error', function( err ) { // Handle errors
                 fs.unlink( dest ); // Delete the file async. (But we don't check the result)
-                console.log( '[downloadJQuery] ERROR: ' + err.message );
+                console.log( '[downloadJQueryMin] ERROR: ' + err.message );
                 done();
             });
         },
@@ -146,23 +134,61 @@ var TzeGenerator = yeoman.generators.Base.extend({
         downloadJQueryMinMap: function () {
             var done = this.async();
             var jqVrsn = this.jqVer;
-            var dest = 'src/js/libs/jquery.min.map';
-            var file = fs.createWriteStream( dest );
             var skipMsg = this.options['skip-message'];
-            var request = http.get('http://cdnjs.cloudflare.com/ajax/libs/jquery/' + jqVrsn + '/jquery.min.map', function(response) {
-                response.pipe( file ).on( 'close', function () {
-                     postdata( fs.readFileSync( dest ) );
-                    if (!skipMsg) {
-                        console.log('[downloadJQueryMinMap] jQuery.min.map downloaded.');
-                    }
+
+            if ( jqVrsn != '1.8.3' && jqVrsn != '1.7.2' ) {
+                var dest = 'src/js/libs/jquery.min.map';
+                var file = fs.createWriteStream( dest );
+                var request = http.get( 'http://cdnjs.cloudflare.com/ajax/libs/jquery/' + jqVrsn + '/jquery.min.map', function( response ) {
+                    response.pipe( file ).on( 'close', function () {
+                         postdata( fs.readFileSync( dest ) );
+                        if ( !skipMsg ) {
+                            console.log( '[downloadJQueryMinMap] jQuery.min.map downloaded.' );
+                        }
+                        done();
+                    });
+                })
+                .on( 'error', function( err ) { // Handle errors
+                    fs.unlink( dest ); // Delete the file async. (But we don't check the result)
+                    console.log( '[downloadJQueryMinMap] ERROR: ' + err.message );
                     done();
                 });
-            })
-            .on( 'error', function( err ) { // Handle errors
-                fs.unlink( dest ); // Delete the file async. (But we don't check the result)
-                console.log( '[downloadJQuery] ERROR: ' + err.message );
-                done();
-            });
+            } else {
+                if (!skipMsg) {
+                    console.log( '[downloadJQueryMinMap] No map downloaded -- version lower than 1.9.' );
+                    done();
+                }
+            }
+        },
+
+        downloadJQuery: function () {
+            var done = this.async();
+            var jqVrsn = this.jqVer;
+            var skipMsg = this.options['skip-message'];
+
+            if ( jqVrsn != '1.8.3' && jqVrsn != '1.7.2' ) {
+                var dest = 'src/js/libs/jquery.js';
+                var file = fs.createWriteStream( dest );
+                var request = http.get( 'http://cdnjs.cloudflare.com/ajax/libs/jquery/' + jqVrsn + '/jquery.js', function( response ) {
+                    response.pipe( file ).on( 'close', function () {
+                         postdata( fs.readFileSync( dest ) );
+                        if ( !skipMsg ) {
+                            console.log( '[downloadJQuery] Unminified jQuery downloaded: ' + jqVrsn );
+                        }
+                        done();
+                    });
+                })
+                .on( 'error', function( err ) { // Handle errors
+                    fs.unlink( dest ); // Delete the file async. (But we don't check the result)
+                    console.log( '[downloadJQuery] ERROR: ' + err.message );
+                    done();
+                });
+            } else {
+                if (!skipMsg) {
+                    console.log( '[downloadJQuery] No unminified jQuery downloaded -- version lower than 1.9.' );
+                    done();
+                }
+            }
         },
 
         generatePackageJson: function () {
@@ -180,11 +206,12 @@ var TzeGenerator = yeoman.generators.Base.extend({
             }
         },
 
-        copyMainFiles: function () {
+        copyFiles: function () {
             this.copy("_main.scss", "src/sass/main.scss");
             this.copy("_reset.scss", "src/sass/libs/_reset.scss");
-            this.copy("_jquery-ui-1.11.2.min.css", "src/sass/libs/_jquery-ui-1.11.2.min.scss");
-            this.copy("_jquery-ui.structure-1.11.2.min.css", "src/sass/libs/_jquery-ui.structure-1.11.2.min.scss");
+            this.copy("_jquery-ui-1.11.4.min.css", "src/sass/libs/_jquery-ui-1.11.4.min.scss");
+            this.copy("_jquery-ui.structure-1.11.4.min.css", "src/sass/libs/_jquery-ui.structure-1.11.4.min.scss");
+            this.copy("_jquery-ui.theme-1.11.4.min.css", "src/sass/libs/_jquery-ui.theme-1.11.4.min.scss");
             this.copy("_all.scss", "src/sass/modules/_all.scss");
             this.copy("_vars.scss", "src/sass/modules/_vars.scss");
             this.copy("_mixins.scss", "src/sass/modules/_mixins.scss");
@@ -195,11 +222,13 @@ var TzeGenerator = yeoman.generators.Base.extend({
             this.copy("_masthead.scss", "src/sass/partials/_masthead.scss");
             this.copy("_component1.scss", "src/sass/partials/_component1.scss");
             this.copy("_component2.scss", "src/sass/partials/_component2.scss");
-            this.copy("_jquery-ui-1.11.2.min.js", "src/js/libs/jquery-ui-1.11.2.min.js");
+            this.copy("_jquery-ui-1.11.4.min.js", "src/js/libs/jquery-ui-1.11.4.min.js");
             this.copy("_modernizr.2.8.3.custom.js", "src/js/libs/modernizr.2.8.3.custom.js");
 
+            this.directory("images", "src/css/images");
+
             if (!this.options['skip-message']) {
-                console.log('[copyMainFiles] Main files generated.');
+                console.log('[copyFiles] Main files generated.');
             }
         },
